move more things to utils
