--- packages/cli/docker-compose.yaml
@@ -1,27 +1,0 @@
-services:
-  db:
-    image: postgres:13
-    container_name: db
-    ports:
-      - '5432:5432'
-    environment:
-      POSTGRES_USER: ${POSTGRES_USER:-postgres}
-      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-postgres}
-      POSTGRES_DB: ${POSTGRES_DB:-postgres}
-    volumes:
-      - ./db:/var/lib/postgresql/data
-
-  inngest:
-    build:
-      context: ./inngest
-      dockerfile: Dockerfile
-    container_name: inngest
-    image: inngest:latest
-    ports:
-      - '8288:8288'
-    extra_hosts:
-      - 'host.docker.internal:host-gateway'
-    command:
-      - '--no-discovery'
-      - '-u'
-      - ${INNGEST_SERVE_URL:-http://host.docker.internal:3000/api/integrations/inngest}
