+++ packages/cli/src/commands/migrate.ts
@@ -1,44 +1,5 @@
 import { execa, ExecaError } from 'execa';
-import path from 'path';
-
-async function checkPostgresReady(dbUrl: string) {
-  for (let i = 0; i < 10; i++) {
-    try {
-      await _migrate(true, dbUrl, true); // attempts to create the migration w/o applying it
-      return true;
-    } catch (error) {
-      console.log(`Waiting for postgres to be ready, attempt ${i + 1} of 10`);
-    }
-    await new Promise(resolve => setTimeout(resolve, 1000));
-  }
-  throw new Error('Postgres is not ready, aborting');
-}
-
-async function _migrate(createOnly = false, dbUrl: string, swallow: boolean = false) {
-  const PRISMA_BIN = path.resolve(
-    process.cwd(),
-    'node_modules',
-    '@arkw/core',
-    'node_modules',
-    'prisma',
-    'node_modules',
-    '.bin',
-  );
-
-  const PRISMA_SCHEMA = path.resolve(process.cwd(), 'node_modules', '@arkw/core', 'src', 'prisma', 'schema.prisma');
-
-  const CREATE_ONLY = createOnly ? `--create-only` : ``;
-
-  return execa(`${PRISMA_BIN}/prisma migrate dev ${CREATE_ONLY} --schema=${PRISMA_SCHEMA} --name initial_migration`, {
-    env: {
-      ...process.env,
-      DB_URL: dbUrl,
-    },
-    shell: true,
-    all: true,
-    stdio: swallow ? 'ignore' : 'inherit', // inherit will pipe directly to parent process stdout/stderr
-  });
-}
+import { getPrismaBinPath, getPrismaSchemaPath } from '../utils.js';
 
 export async function migrate(createOnly = false, dbUrl: string) {
   console.log('Migrating database...');
@@ -58,4 +19,35 @@ export async function migrate(createOnly = false, dbUrl: string) {
     }
   }
   return false;
+}
+
+export async function _migrate(createOnly = false, dbUrl: string, swallow: boolean = false) {
+  const PRISMA_BIN = getPrismaBinPath();
+
+  const PRISMA_SCHEMA = getPrismaSchemaPath();
+
+  const CREATE_ONLY = createOnly ? `--create-only` : ``;
+
+  return execa(`${PRISMA_BIN} migrate dev ${CREATE_ONLY} --schema=${PRISMA_SCHEMA} --name initial_migration`, {
+    env: {
+      ...process.env,
+      DB_URL: dbUrl,
+    },
+    shell: true,
+    all: true,
+    stdio: swallow ? 'ignore' : 'inherit', // inherit will pipe directly to parent process stdout/stderr
+  });
+}
+
+async function checkPostgresReady(dbUrl: string) {
+  for (let i = 0; i < 10; i++) {
+    try {
+      await _migrate(true, dbUrl, true); // attempts to create the migration w/o applying it
+      return true;
+    } catch (error) {
+      console.log(`Waiting for postgres to be ready, attempt ${i + 1} of 10`);
+    }
+    await new Promise(resolve => setTimeout(resolve, 1000));
+  }
+  throw new Error('Postgres is not ready, aborting');
 }
