+++ packages/cli/src/commands/provision.ts
@@ -1,24 +1,24 @@
 import { execa } from 'execa';
-import fs from 'fs';
 import path from 'path';
 import prompt from 'prompt';
-import { check } from 'tcp-port-used';
 
+import fs from 'fs';
 import fse from 'fs-extra/esm';
 
-import { config as defaultConfig } from '../files/starter-config.js';
+import { config as defaultConfig } from '../starter-files/config.js';
 import { FileEnvService } from '../services/service.fileEnv.js';
-import { replaceValuesInFile } from '../utils.js';
+import { replaceValuesInFile, copyStarterFile, sanitizeForDockerName, getInfraPorts } from '../utils.js';
 
-const DOCKER_COMPOSE_FILENAME = 'arkw.docker-compose.yaml';
+const DOCKER_COMPOSE_FILE = 'arkw.docker-compose.yaml';
+const ARKW_CONFIG_FILE = 'arkw.config.ts';
 
 export async function provision(projectName: string) {
   const sanitizedProjectName = sanitizeForDockerName(projectName);
 
   const { postgresPort, inngestPort } = await getInfraPorts();
 
-  await setupConfigFile({ postgresPort, sanitizedProjectName });
-  await setupRoutesFile();
+  await setupConfig({ postgresPort, sanitizedProjectName });
+  await setupRoutes();
 
   const { userInputDbUrl, userInputInngestUrl } = await promptUserForInfra();
 
@@ -33,9 +33,12 @@ export async function provision(projectName: string) {
   });
 
   if (shouldRunDocker) {
-    console.log('Starting Docker containers...');
     try {
-      await execa('docker', ['compose', '-f', DOCKER_COMPOSE_FILENAME, 'up', '-d'], { stdio: 'inherit' });
+      await execa(
+        'docker', 
+        ['compose', '-f', DOCKER_COMPOSE_FILE, 'up', '-d'], 
+        { stdio: 'inherit' }
+      );
       console.log('Docker containers started successfully.');
     } catch (error) {
       console.error('Failed to start Docker containers:', error);
@@ -59,12 +62,12 @@ export function prepareDockerComposeFile({
   postgresPort: number;
   inngestPort: number;
 }) {
-  let inngestUrl: string;
+  let inngestUrl = `http://localhost:${inngestPort}`;
   let dbUrl = `postgresql://postgres:postgres@localhost:${postgresPort}/arkwright?schema=arkw`;
 
   const editDockerComposeFile = () => {
     replaceValuesInFile({
-      filePath: DOCKER_COMPOSE_FILENAME,
+      filePath: DOCKER_COMPOSE_FILE,
       replacements: [
         { replace: sanitizedProjectName, search: 'REPLACE_PROJECT_NAME' },
         { replace: `${postgresPort}`, search: 'REPLACE_DB_PORT' },
@@ -74,13 +77,10 @@ export function prepareDockerComposeFile({
   };
 
   if (userInputDbUrl === '' && userInputInngestUrl === '') {
-    console.log('Creating new PostgreSQL instance and Inngest server...');
-    copyStarterFile('starter-docker-compose.yaml', DOCKER_COMPOSE_FILENAME);
+    copyStarterFile('docker-compose-pg-inngest.yaml', DOCKER_COMPOSE_FILE);
     editDockerComposeFile();
-    inngestUrl = `http://localhost:${inngestPort}`;
   } else if (userInputDbUrl === '' && userInputInngestUrl !== '') {
-    console.log('Setting up new Inngest server...');
-    copyStarterFile('starter-docker-compose-postgres.yaml', DOCKER_COMPOSE_FILENAME);
+    copyStarterFile('docker-compose-pg-only.yaml', DOCKER_COMPOSE_FILE);
     editDockerComposeFile();
     inngestUrl = String(userInputInngestUrl);
   } else if (userInputDbUrl !== '' && userInputInngestUrl === '') {
@@ -93,23 +93,7 @@ export function prepareDockerComposeFile({
   return { dbUrl, inngestUrl };
 }
 
-export function copyStarterFile(inputFile: string, outputFile: string) {
-  const __filename = new URL(import.meta.url).pathname;
-  const __dirname = path.dirname(__filename);
-  const filePath = path.resolve(__dirname, '..', '..', 'src', 'files', inputFile);
-  const fileString = fs.readFileSync(filePath, 'utf8');
-
-  const outputFilePath = path.join(process.cwd(), outputFile);
-  if (fs.existsSync(outputFilePath)) {
-    console.log(`${outputFile} already exists`);
-    return false;
-  }
-
-  fse.outputFileSync(outputFilePath, fileString);
-  return fileString;
-}
-
-export async function setupRoutesFile() {
+export async function setupRoutes() {
   const { routeRegistrationPath } = defaultConfig;
   const tsconfigPath = path.resolve(process.cwd(), 'tsconfig.json');
   const tsconfig = JSON.parse(fs.readFileSync(tsconfigPath, 'utf8'));
@@ -127,7 +111,7 @@ export async function setupRoutesFile() {
     return;
   }
 
-  copyStarterFile('starter-api.ts', apiPath);
+  copyStarterFile('api.ts', apiPath);
 }
 
 async function promptUserForInfra() {
@@ -155,17 +139,17 @@ async function promptUserForInfra() {
   return { userInputDbUrl, userInputInngestUrl };
 }
 
-async function setupConfigFile({
+async function setupConfig({
   postgresPort,
   sanitizedProjectName,
 }: {
   postgresPort: number;
   sanitizedProjectName: string;
 }) {
-  copyStarterFile('starter-config.ts', 'arkw.config.ts');
+  copyStarterFile('config.ts', ARKW_CONFIG_FILE);
 
   replaceValuesInFile({
-    filePath: 'arkw.config.ts',
+    filePath: ARKW_CONFIG_FILE,
     replacements: [
       {
         search: 'REPLACE_DB_PORT',
@@ -188,60 +172,4 @@ export async function setupEnvFile({ inngestUrl, dbUrl }: { inngestUrl: string; 
   await fileEnvService.setEnvValue('INNGEST_URL', inngestUrl);
   await fileEnvService.setEnvValue('DB_URL', dbUrl);
   await fileEnvService.setEnvValue('APP_URL', 'http://localhost:3000');
-}
-
-const isPortOpen = async (port: number): Promise<boolean> => {
-  return new Promise((resolve, reject) => {
-    check(port).then((inUse: boolean) => {
-      resolve(!inUse);
-    });
-  });
-};
-
-const getNextOpenPort = async (startFrom: number = 2222): Promise<number> => {
-  for (const port of Array.from({ length: 20 }, (_, i) => startFrom + i)) {
-    const isOpen = await isPortOpen(port);
-    if (isOpen) {
-      return port;
-    }
-  }
-  throw new Error('No open ports found after 20 attempts');
-};
-
-async function getInfraPorts() {
-  let postgresPort = 5432;
-  let inngestPort = 8288;
-  const dbPortOpen = await isPortOpen(postgresPort);
-  const inngestPortOpen = await isPortOpen(inngestPort);
-
-  if (!dbPortOpen) {
-    postgresPort = (await getNextOpenPort(postgresPort)) as number;
-  }
-
-  if (!inngestPortOpen) {
-    inngestPort = (await getNextOpenPort(inngestPort)) as number;
-  }
-
-  return { postgresPort, inngestPort };
-}
-
-function sanitizeForDockerName(name: string): string {
-  // Convert to lowercase
-  let sanitized = name.toLowerCase();
-
-  // Replace any non-alphanumeric characters (excluding dashes) with dashes
-  sanitized = sanitized.replace(/[^a-z0-9-]/g, '-');
-
-  // Trim dashes from the start and end
-  sanitized = sanitized.replace(/^-+|-+$/g, '');
-
-  // Ensure name is between 2 and 255 characters
-  if (sanitized.length < 2) {
-    throw new Error('Name must be at least 2 characters long.');
-  }
-  if (sanitized.length > 255) {
-    sanitized = sanitized.substring(0, 255);
-  }
-
-  return sanitized;
 }
