+++ packages/cli/src/starter-files/config.ts
